// frontend/client/src/types.tsx
export interface APIResponse<Results> {
  error: APIError | Error | null;
  isLoading: boolean;
  response: ParsedResponse<Results> | null;
}
export type Permissions = {
  [key in typeof RESOURCES[number]]?: typeof PERMISSIONS[number][];
};
export type APIError = { status: number; moreInfo: string; message: string };